% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_button_leaflet.R
\name{add_bc_home_button}
\alias{add_bc_home_button}
\title{Add a 'Zoom to BC button' to a leaflet map}
\usage{
add_bc_home_button(map, ...)
}
\arguments{
\item{map}{a Leaflet map object}

\item{...}{Other arguments passed on to `leaflet::easyButton()`, such 
as `position` (default is `"topleft"`)}
}
\value{
a Leaflet map object
}
\description{
Add a 'Zoom to BC button' to a leaflet map
}
