% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{css_caaqs_copy}
\alias{css_caaqs_copy}
\title{Create copy of CAAQS CSS styles for leaflet map}
\usage{
css_caaqs_copy(folder = "./leaflet_map", overwrite = FALSE)
}
\arguments{
\item{folder}{Character. Location of the leaflet maps folder.}

\item{overwrite}{Logical. Overwrite the CSS file if it already exits?}
}
\description{
Creates a copy in the local repository of the CAAQS CSS file.
}
