% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{envreportutils-deprecated}
\alias{envreportutils-deprecated}
\alias{roxygen_template}
\alias{devex_badge}
\alias{analysis_skeleton}
\alias{add_code_of_conduct}
\alias{add_contributing}
\alias{add_license}
\alias{add_license_header}
\alias{add_readme}
\alias{add_rproj}
\title{Defunct functions in envreportutils}
\usage{
roxygen_template(...)

devex_badge(...)

analysis_skeleton(...)

add_code_of_conduct(...)

add_contributing(...)

add_license(...)

add_license_header(...)

add_readme(...)

add_rproj(...)
}
\arguments{
\item{...}{any paramters passed are ignored}
}
\description{
These functions are no longer available in this package.
}
\details{
All of these functions, except roxygen_template, are available in the bcgovr
package available on GitHub here: https://github.com/bcgov/bcgovr

\itemize{
 \item \code{\link{roxygen_template}}: This function is defunct.
  \item \code{\link{devex_badge}}: This function is defunct.
   \item \code{\link{analysis_skeleton}}: This function is defunct.
   \item \code{\link{add_code_of_conduct}}: This function is defunct.
   \item \code{\link{add_contributing}}: This function is defunct.
    \item \code{\link{add_license}}: This function is defunct.
   \item \code{\link{add_license_header}}: This function is defunct.
   \item \code{\link{add_readme}}: This function is defunct.
   \item \code{\link{add_rproj}}: This function is defunct.
}
}
