% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envreportutils-package.R
\docType{package}
\name{envreportutils}
\alias{envreportutils}
\title{envreportutils}
\description{
A package with ggplot2 themes and other functions commonly used by the Environmental Reporting BC
team when developing environmental reporting indicators and related products.
}
