% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_licence.R
\name{get_data_licence}
\alias{get_data_licence}
\alias{get_data_license}
\title{Returns a url, or a markdown-formatted or html-formatted link to one of several B.C. or Canadian licences}
\usage{
get_data_licence(licence = c("ogl-bc", "bc-crown", "ogl-canada",
  "statscan"), what = "md_link")
}
\arguments{
\item{licence}{Which licence? One of:
\itemize{
  \item ogl-bc Open Government Licence - British Columbia
  \item bc-crown B.C. Crown Copyright
  \item ogl-canada Open Government Licence - Canada
  \item statscan Statistics Canada Open Licence Agreement
}}

\item{what}{What do you want returned? One of
\itemize{
  \item md_link A markdown-formatted link (default)
  \item html_link An HTML-formatted link
  \item url The url of the licence
}}
}
\value{
Character containing the desired output
}
\description{
Returns a url, or a markdown-formatted or html-formatted link to one of several B.C. or Canadian licences
}
\examples{
get_data_licence("ogl-bc", "url")

get_data_licence("ogl-canada")

get_data_licence("statscan", "html_link")
}
