% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percent_round}
\alias{percent_round}
\title{Round percentage values in standard way for reporting}
\usage{
percent_round(x, round_90_to_one = FALSE)
}
\arguments{
\item{x}{numerice vector of percentages}

\item{round_90_to_one}{Should values above 90% be reported to a single 
decimal place?}
}
\value{
character vector of formatted percentages following these rounding rules
}
\description{
Follows the advice from: http://adc.bmj.com/content/early/2015/04/15/archdischild-2014-307149
"Percentages: [Use] integers, or one decimal place for values under 10%. 
Values over 90% may need one decimal place if their complement is informative. 
Use two or more decimal places only if the range of values is less than 0.1%"
}
\examples{
percent_round(c(2.5, 11.324, 95.898, 99.6))
percent_round(c(2.5, 11.324, 95.898, 99.6), round_90_to_one = TRUE)
}
