% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_devices.R
\name{png_retina}
\alias{png_retina}
\alias{save_png_retina}
\title{Create png for retina display}
\usage{
png_retina(
  filename = "Rplot\%03d.png",
  width = 480,
  height = 480,
  units = "px",
  pointsize = 12,
  bg = "white",
  res = NA,
  ...,
  type = c("cairo", "cairo-png", "Xlib", "quartz"),
  antialias
)

save_png_retina(
  x,
  filename = "Rplot\%03d.png",
  width = 480,
  height = 480,
  units = "px",
  pointsize = 12,
  bg = "white",
  res = NA,
  ...,
  type = c("cairo", "cairo-png", "Xlib", "quartz"),
  antialias
)
}
\arguments{
\item{filename}{the path of the output file, up to 511 characters.  The
    page number is substituted if a C integer format is included in the
    character string, as in the default, and tilde-expansion is
    performed (see \code{\link{path.expand}}).   (The result must be less than
    600 characters long.  See \code{\link[grDevices]{postscript}} for further
    details.)
  }

\item{width}{the width of the device.}

\item{height}{the height of the device.}

\item{units}{The units in which \code{height} and \code{width} are
    given.  Can be \code{px} (pixels, the default), \code{in} (inches),
    \code{cm} or \code{mm}.}

\item{pointsize}{the default pointsize of plotted text, interpreted as
    big points (1/72 inch) at \code{res} ppi.}

\item{bg}{the initial background colour: can be overridden by setting
    par("bg").}

\item{res}{The nominal resolution in ppi which will be recorded in the
    bitmap file, if a positive integer.  Also used for units other than
    the default.  If not specified, taken as 72 ppi to set the size of
    text and line widths.}

\item{type}{Should be plotting be done using Windows GDI or cairographics?}

\item{antialias}{Length-one character vector.

    For allowed values and their effect on fonts with \code{type =
      "windows"} see \code{\link[grDevices]{windows}}: for that type if the
    argument is missing the default is taken from
    \code{windows.options()$bitmap.aa.win}.

    For allowed values and their effect (on fonts and lines, but not
    fills) with \code{type = "cairo"} see \code{\link[grDevices]{svg}}.}

\item{x}{a ggplot2 object}
}
\value{
A plot device is opened: nothing is returned to the R interpreter
}
\description{
`png_retina` is a drop-in replacement for the \code{\link[grDevices]{png}} function 
for creating images for the web for retina devices. Internally, it simply 
doubles the width, height, and resolution specified. The intention is then
that in the webpage that you would specify the \code{width} and \code{height} 
attributes in the html at the original resolution.

`save_png_retina` is a wrapper function around `png_retina` analagous to 
[ggplot2::gsave()]
}
\section{Functions}{
\itemize{
\item \code{save_png_retina()}: 

}}
\examples{

# You want to display at 500 * 500 in the web:
png_retina("myplot.svg", width = 500, height = 500)
plot(x = 1:500, y = rnorm(500))
dev.off()

# Although the output image will be 1000 * 1000, in the html you would put:
# <img src="myplot.png", width="500", height="500" />

if (suppressPackageStartupMessages(require("ggplot2", quietly = TRUE))) {
p <- ggplot(mtcars, aes(x = cyl, y = mpg)) + geom_point()
file <- tempfile(fileext = ".png")
save_png_retina(p, file)
}
}
\seealso{
\code{\link[grDevices]{png}}
}
