% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{popup_caaqs}
\alias{popup_caaqs}
\title{Create popup for CAAQS indicators}
\usage{
popup_caaqs(data, type = "station", metric_name, units, standard_name)
}
\arguments{
\item{data}{Data Frame. CAAQ information}

\item{type}{Character. Which type of popup? "station" or "region"?}

\item{metric_name}{Character. Display name of the CAAQ metric in HTML code
(e.g., "Ozone Metric", or "PM<sub>2.5</sub> Metric (annual)").}

\item{units}{Character. Metric units in HTML code (e.g., "ppm" or
"&mu;g/m&sup3;")}

\item{standard_name}{Character. Display name of the CAAQ standard in HTML
code (e.g., "Ozone Air Quality Standard" or "PM<sub>2.5</sub> Air Quality
Standard (annual)")}
}
\value{
Character vector of the HTML code for the popup to be passed to
  leaflet
}
\description{
Create popup for CAAQS indicators
}
\details{
Data frame must contain the following columns: 'p_az' reflecting the
  airzone name, 'p_station_id' reflecting the station id, 'p_station'
  reflecting the station name, 'ambient_metric_value' reflecting the CAAQS metric
  value, 'ambient_caaqs' reflecting the CAAQS status (not in HTML code), and
  'n_years' reflecting the number of years the CAAQS metric is averaged over.
}
