% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{popup_caaqs_combo}
\alias{popup_caaqs_combo}
\title{Create popup for combination CAAQS indicators}
\usage{
popup_caaqs_combo(data, type = "station", metric_type, metrics, units)
}
\arguments{
\item{data}{Data Frame. CAAQ information}

\item{type}{Character. Which type of popup? "station" or "region"?}

\item{metric_type}{Character Display name of the CAAQ metric in HTML code
(e.g., `"PM<sub>2.5</sub>"`).}

\item{metrics}{Character vector. Metrics to combine (e.g., c("annual",
"24hr"))}

\item{units}{Character. Metric units in HTML code (e.g., "ppm" or
"&mu;g/m&sup3;")}
}
\value{
Character vector of the HTML code for the popup to be passed to
  leaflet
}
\description{
For CAAQs where two metrics used to assess management status (e.g., 
PM 2.5, SO2, NO2)
}
\details{
Data frame must contain the following columns: 'p_az' reflecting the
  airzone name, 'p_station_id' reflecting the station id, 'p_station'
  reflecting the station name, 'colour' reflecting the background colour for
  the CAAQS achievement box, 'metric_value_ambient' reflecting the CAAQS
  metric value, 'caaqs_ambient' reflecting the CAAQS status (not in HTML
  code), and 'n_years' reflecting the number of years the CAAQS metric is
  averaged over.
}
