% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{popup_combine_rows}
\alias{popup_combine_rows}
\title{Combine popup rows for leaflet maps}
\usage{
popup_combine_rows(data)
}
\arguments{
\item{Character.}{Row content}
}
\description{
Combine popup rows for leaflet maps
}
