% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{popup_create_row}
\alias{popup_create_row}
\title{Create a popup row div for leaflet maps}
\usage{
popup_create_row(...)
}
\arguments{
\item{...}{Character. Row content}
}
\description{
Create a popup row div for leaflet maps
}
