% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{popup_groundwater}
\alias{popup_groundwater}
\title{Create content of leaflet popups for groundwater levels}
\usage{
popup_groundwater(data, type = "well")
}
\arguments{
\item{data}{Data frame. Contains information relevant for popup (see Details).}

\item{type}{Character. Which type of popup to create? "well" or "region"}
}
\description{
This is a helper function to format popups containing figures and information
for wells in the groundwater-levels-indicator
}
\details{
For the svg plots, <code>data</code> must contain columns
  <code>well_num</code> and <code>region_name_short</code> which reflect the
  svg plot names. For the information content, <code>data</code> must also
  contain columns <code>region_name</code>, <code>well_num</code>, and
  <code>state</code> which reflect the name of the region, the observation
  well number and the trend state, respectively.
}
