% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{popups_centre}
\alias{popups_centre}
\title{Keep popups centred in view port}
\usage{
popups_centre(map)
}
\arguments{
\item{map}{Map. A Leaflet map object}
}
\value{
A Leaflet map object
}
\description{
Reloads popups the first time they're opened to correctly pan popup in
display. Based on https://stackoverflow.com/a/38172374, EDIT 2 (example:
https://jsfiddle.net/3v7hd2vx/277/)
}
