% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_percent}
\alias{report_percent}
\title{Round percentage values in standard way for reporting}
\usage{
report_percent(x, round_90_to_one = FALSE, as_char = TRUE)
}
\arguments{
\item{x}{Numeric vector of percentages.}

\item{round_90_to_one}{Should values above 90 percent be reported to a single 
decimal place?}

\item{as_char}{Should the result be returned as a character vector? Default `TRUE`.}
}
\value{
Character vector of formatted percentages following these rounding rules
}
\description{
Follows the advice from: http://adc.bmj.com/content/early/2015/04/15/archdischild-2014-307149
"Percentages: [Use] integers, or one decimal place for values under 10%. 
Values over 90% may need one decimal place if their complement is informative. 
Use two or more decimal places only if the range of values is less than 0.1%"
}
\examples{
report_percent(c(2.5, 11.324, 95.898, 99.6))
report_percent(c(2.5, 11.324, 95.898, 99.6), round_90_to_one = TRUE)
}
