% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_button_leaflet.R
\name{set_bc_view}
\alias{set_bc_view}
\title{Set Leaflet map view to B.C.}
\usage{
set_bc_view(map, zoom = 5)
}
\arguments{
\item{map}{a Leaflet map object}

\item{zoom}{Zoom level, default `5`}
}
\value{
a Leaflet map object
}
\description{
Set Leaflet map view to B.C.
}
