% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{set_bc_view_on_close}
\alias{set_bc_view_on_close}
\title{Re-centre map to B.C. on popup close}
\usage{
set_bc_view_on_close(map, zoom = 5)
}
\arguments{
\item{map}{Map. A Leaflet map object}

\item{zoom}{Numeric. Zoom level, default `5`}
}
\value{
A Leaflet map object
}
\description{
Re-centre map to B.C. on popup close
}
