% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_soe_root}
\alias{set_soe_root}
\title{Set the path to the root of your SoE network folder}
\usage{
set_soe_root(x)
}
\arguments{
\item{x}{character. A path to the SOE folder on the network (pickaxe) drive.
For Mac users it will look something like `"/Users/username/Volumes/pickaxe/SOE"`. 
For Windows users, use the mapped drive letter rather than the UNC path 
(e.g., `"P:/pickaxe/SOE"`)}
}
\value{
logical (invisibly)
}
\description{
This will add an environment variable called `ENVREPORTUTILS_SOE_PATH` to 
your `.Renviron` file and set it to the path on your computer to the 
`pickaxe/SOE` folder.
}
\examples{
\dontrun{
# Mac
set_soe_root("/Users/username/Volumes/pickaxe/SOE")
# Windows
set_soe_root("P:/pickaxe/SOE")
}
}
