% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{soe_path}
\alias{soe_path}
\title{Specify a path to a folder or file in the SOE folder on the network drive}
\usage{
soe_path(x)
}
\arguments{
\item{x}{character. Path to a file or folder, relative to `pickaxe/SOE`}
}
\value{
full path to the file or folder
}
\description{
Specify a path to a folder or file in the SOE folder on the network drive
}
\examples{
\dontrun{
soe_path("Operations ORCS/indicators/air")
# In a function call:
read.csv(soe_path("Operations ORCS/data/foo.csv"))
}
}
