% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft.R
\name{soft}
\alias{soft}
\title{Post a file to SOFT for sharing}
\usage{
soft(file, email = NULL, email_subj = NULL, internal = TRUE,
  days = 7, progress = TRUE, zip = FALSE, zipname = NULL)
}
\arguments{
\item{file}{Path to the file or directory on your computer. If \code{file} is a
directory or a character vector of filenames, the files in the directory will
be zipped before they are uploaded.}

\item{email}{Optional email address to which to send the link to the file.}

\item{email_subj}{Optional subject of the email.}

\item{internal}{Should the link be only available with the B.C. Government
network? Default \code{TRUE}}

\item{days}{The number of days the file should be available.
Default 7, maximum 14.}

\item{progress}{Should a progress bar be displayed? Default \code{TRUE}.}

\item{zip}{Should the file be zipped before sending? Default \code{FALSE},
however it \code{file} is a directory or multiple files they will be zipped
anyway.}

\item{zipname}{an option name to call the zipfile. If left \code{NULL} (default),
a name will be generated.}
}
\value{
A url of the link from which the file can be downloaded
}
\description{
Files can be shared from within the government network, from a government user to the public, or
from the public to a government user. Public-to-public sharing is not supported. A user can specify the
number of days the file will be available, and whether or not to restrict access to the file to
within the governmnet network.
}
\details{
The SOFT web form and more information can be found \href{http://www.env.gov.bc.ca/csd/imb/soft/}{here}.
}
