% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_devices.R
\name{svg_px}
\alias{svg_px}
\title{Create svg for the web, specifying size in pixels}
\usage{
svg_px(file = "Rplots.svg", width = 800, height = 800, res = 72,
  bg = "white", pointsize = 12, standalone = TRUE,
  system_fonts = list(), user_fonts = list())
}
\arguments{
\item{width}{The width of the plot in pixels. Default \code{800}.}

\item{height}{The height of the plot in pixels. Default \code{800} .}

\item{res}{The resolution at which it is displayed. Default \code{72}. 
You should rarely have to change this.}
}
\value{
A plot device is opened: nothing is returned to the R interpreter
}
\description{
This is a thin wrapper around \code{\link[svglite]{svglite}} 
for creating images for the web in svg format, but enables you specify the 
size in pixels, which is more useful for web work. Internally, it simply 
converts the width and height to inches at the resolution specified.
}
\examples{

# You want to display at 500 * 500 in the web:
svg_px("myplot.png", width = 500, height = 500)
plot(x = 1:500, y = rnorm(500))
dev.off()
}
\seealso{
\code{\link[svglite]{svglite}}
}
