% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_soe.r
\name{theme_bc_sans}
\alias{theme_bc_sans}
\title{BC sans theme for EnvReportBC graphs and plots}
\usage{
theme_bc_sans(base_size = 12, base_family = "BCSans")
}
\arguments{
\item{base_size}{base font size (default = 12)}

\item{base_family}{base font family (default = BC Sans)}
}
\value{
returns a plot theme
}
\description{
Function applies BC sans font to standard SOE theme.
}
\keyword{plotting}
\keyword{theme}
