% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_titlecase}
\alias{to_titlecase}
\title{Convert a word or phrase to Title Case}
\usage{
to_titlecase(x)
}
\arguments{
\item{x}{character string}
}
\value{
A character string: x in Title Case
}
\description{
Convert a word or phrase to Title Case
}
\examples{
x <- "hello world"
to_titlecase(x)

x <- "HELLO WORLD"
to_titlecase(x)

x <- "heLLo WoRLd"
to_titlecase(x)
}
