% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_devices.R
\name{svg_px}
\alias{svg_px}
\title{Create svg for the web, specifying size in pixels}
\usage{
svg_px(file = "Rplots.svg", width = 800, height = 800, res = 72,
  bg = "white", pointsize = 12, standalone = TRUE,
  system_fonts = list(), user_fonts = list())
}
\arguments{
\item{file}{The file where output will appear.}

\item{width}{the width of the plot in pixels. Default \code{800}.}

\item{height}{the height of the plot in pixels. Default \code{800}.}

\item{res}{the resolution at which it is displayed. Default \code{72}. 
You should rarely have to change this}

\item{bg}{Default background color for the plot (defaults to "white").}

\item{pointsize}{Default point size.}

\item{standalone}{Produce a standalone svg file? If \code{FALSE}, omits
xml header and default namespace.}

\item{system_fonts}{Named list of font names to be aliased with
fonts installed on your system. If unspecified, the R default
families \code{sans}, \code{serif}, \code{mono} and \code{symbol}
are aliased to the family returned by
\code{\link[gdtools]{match_family}()}.}

\item{user_fonts}{Named list of fonts to be aliased with font files
provided by the user rather than fonts properly installed on the
system. The aliases can be fonts from the fontquiver package,
strings containing a path to a font file, or a list containing
\code{name} and \code{file} elements with \code{name} indicating
the font alias in the SVG output and \code{file} the path to a
font file.}
}
\value{
A plot device is opened: nothing is returned to the R interpreter.
}
\description{
This is a thin wrapper around \code{\link[svglite]{svglite}} 
for creating images for the web in svg format, but enables you specify the 
size in pixels, which is more useful for web work. Internally, it simply 
converts the width and height to inches at the resolution specified.
}
\examples{

# You want to display at 500 * 500 in the web:
svg_px("myplot.png", width = 500, height = 500)
plot(x = 1:500, y = rnorm(500))
dev.off()
}
\seealso{
\code{\link[svglite]{svglite}}
}
