% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_devices.R
\name{svg_px}
\alias{svg_px}
\alias{save_svg_px}
\title{Create svg for the web, specifying size in pixels}
\usage{
svg_px(
  file = "Rplots.svg",
  width = 800,
  height = 800,
  res = 72,
  bg = "white",
  pointsize = 12,
  standalone = TRUE,
  system_fonts = list(),
  user_fonts = list()
)

save_svg_px(
  x,
  file = "Rplots.svg",
  width = 800,
  height = 800,
  res = 72,
  bg = "white",
  pointsize = 12,
  standalone = TRUE,
  system_fonts = list(),
  user_fonts = list()
)
}
\arguments{
\item{file}{Identical to \code{filename}. Provided for backward compatibility.}

\item{width}{The width of the plot in pixels. Default \code{800}.}

\item{height}{The height of the plot in pixels. Default \code{800}.}

\item{res}{The resolution at which it is displayed. Default \code{72}. 
You should rarely have to change this.}

\item{bg}{Default background color for the plot (defaults to "white").}

\item{pointsize}{Default point size.}

\item{standalone}{Produce a standalone svg file? If \code{FALSE}, omits
xml header and default namespace.}

\item{system_fonts}{\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#superseded'><img src='figures/lifecycle-superseded.svg' alt='Superseded lifecycle'></a>}}{\strong{Superseded}} \emph{Consider using
\code{\link[systemfonts:register_font]{systemfonts::register_font()}} instead}. Named list of font
names to be aliased with fonts installed on your system. If unspecified,
the R default families \code{sans}, \code{serif}, \code{mono} and
\code{symbol} are aliased to the family returned by
\code{\link[systemfonts]{font_info}()}.}

\item{user_fonts}{\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#superseded'><img src='figures/lifecycle-superseded.svg' alt='Superseded lifecycle'></a>}}{\strong{Superseded}} \emph{Consider using
\code{\link[systemfonts:register_font]{systemfonts::register_font()}} instead}. Named list of fonts to
be aliased with font files provided by the user rather than fonts properly
installed on the system. The aliases can be fonts from the fontquiver
package, strings containing a path to a font file, or a list containing
\code{name} and \code{file} elements with \code{name} indicating
the font alias in the SVG output and \code{file} the path to a
font file.}

\item{x}{a ggplot2 object}
}
\value{
A plot device is opened: nothing is returned to the R interpreter
}
\description{
This is a thin wrapper around \code{\link[svglite]{svglite}} 
for creating images for the web in svg format, but enables you specify the 
size in pixels, which is more useful for web work. Internally, it simply 
converts the width and height to inches at the resolution specified.

`save_svg_px` is a wrapper function around `svg_px` analagous to 
[ggplot2::gsave()]
}
\section{Functions}{
\itemize{
\item \code{save_svg_px()}: 

}}
\examples{

# You want to display at 500 * 500 in the web:
svg_px("myplot.png", width = 500, height = 500)
plot(x = 1:500, y = rnorm(500))
dev.off()
if (suppressPackageStartupMessages(require("ggplot2", quietly = TRUE))) {
p <- ggplot(mtcars, aes(x = cyl, y = mpg)) + geom_point()
file <- tempfile(fileext = ".svg")
save_svg_px(p, file)
}
}
\seealso{
\code{\link[svglite]{svglite}}
}
